fix some warnings

+++ Project/tbm/TBMVideoProcessor.m
@@ -53,7 +53,7 @@ NSString* const TBMVideoProcessorErrorReason = @"Problem processing video";
 
 - (void)didFinishConvertingToMpeg4 {
     if (self.exportSession.status != AVAssetExportSessionStatusCompleted) {
-        NSString *description = [NSString stringWithFormat:@"export session completed with non complete status: %d  error: %@", self.exportSession.status, self.exportSession.error];
+        NSString *description = [NSString stringWithFormat:@"export session completed with non complete status: %ld  error: %@", (long)self.exportSession.status, self.exportSession.error];
         NSError *error = [self videoProcessorErrorWithMethod:@"didFinishConvertingToMpeg4" description:description];
         [self handleError:error];
         return;
